refactor(channels): keep listener and shared channel contracts vendor-neutral
The listener and generic channel interfaces are meant to expose
channel-agnostic seams so Slack/Telegram/Discord adapters (or external
custom channels) plug in without vendor logic leaking upward. This PR
had leaked Slack into that layer: lifecycle.ts imported
buildSlackModelPickerBlocks and branched on channelId === "slack", and
the generic ChannelModelHandler / sendDirectReply contracts carried a
slackBlocks field.

Now the /model handler returns vendor-neutral data — a
ChannelModelPickerData payload (current model, entries, available and
recent handles) for every channel — and rendering is the adapter's
job: the Slack adapter converts the payload to Block Kit via
buildSlackModelPickerBlocks, which moves with its option helpers from
channels/commands.ts into channels/slack/model-picker-blocks.ts.
Adapters without rich UI just use the text fallback. The slack-syntax
"Model not found" hint moves behind buildChannelModelNotFoundText,
which reuses the existing channel-aware command prefix helper.

listener/* now has no slack references in this PR's diff.
