+++ src/channels/commands.ts
@@ -2,6 +2,7 @@ import type { ListModelsResponseModelEntry } from "@/types/protocol_v2";
 import { getChannelDisplayName } from "./plugin-registry";
 import type {
   ChannelAdapter,
+  ChannelModelPickerData,
   ChannelRoute,
   InboundChannelMessage,
 } from "./types";
@@ -24,12 +25,12 @@ export type ChannelSlashCommandDefinition = {
 export type ChannelSlashCommandHandlerResult = {
   handled: boolean;
   text?: string;
-  slackBlocks?: unknown[];
+  modelPicker?: ChannelModelPickerData;
 };
 
 type ChannelDirectReplyPayload = {
   text: string;
-  slackBlocks?: unknown[];
+  modelPicker?: ChannelModelPickerData;
 };
 
 export type ChannelSlashCommandHandlers = {
@@ -507,7 +508,7 @@ export function buildChannelNewConversationUnavailableMessage(
   return `${displayName} cannot start a new conversation for this chat because no agent is configured.`;
 }
 
-type ChannelModelListEntry = Pick<
+export type ChannelModelListEntry = Pick<
   ListModelsResponseModelEntry,
   | "id"
   | "handle"
@@ -519,11 +520,6 @@ type ChannelModelListEntry = Pick<
 >;
 
 const DEFAULT_CHANNEL_MODEL_LIST_LIMIT = 8;
-const SLACK_MODEL_PICKER_OPTION_LIMIT = 100;
-const SLACK_MODEL_OPTION_TEXT_LIMIT = 75;
-const SLACK_MODEL_OPTION_VALUE_LIMIT = 75;
-
-export const SLACK_MODEL_SELECT_ACTION_ID = "letta_channel_model_select";
 
 function getModelEntryRank(entry: ChannelModelListEntry): number {
   if (entry.isDefault) return 0;
@@ -545,7 +541,7 @@ function preferModelEntry(
     : current;
 }
 
-function buildModelEntriesByHandle(
+export function buildModelEntriesByHandle(
   entries: ChannelModelListEntry[],
 ): Map<string, ChannelModelListEntry> {
   const byHandle = new Map<string, ChannelModelListEntry>();
@@ -568,7 +564,7 @@ function makeUnknownModelEntry(handle: string): ChannelModelListEntry {
   };
 }
 
-function resolveModelHandles(params: {
+export function resolveModelHandles(params: {
   handles: string[];
   byHandle: Map<string, ChannelModelListEntry>;
   availableHandles?: Set<string> | null;
@@ -585,7 +581,7 @@ function resolveModelHandles(params: {
   return resolved;
 }
 
-function getFallbackModelEntries(
+export function getFallbackModelEntries(
   byHandle: Map<string, ChannelModelListEntry>,
 ): ChannelModelListEntry[] {
   const preferred = Array.from(byHandle.values()).filter(
@@ -596,6 +592,10 @@ function getFallbackModelEntries(
 
 function modelCommandPrefix(channelId: string): "/model" | "@agent /model" {
   return channelId === "slack" ? "@agent /model" : "/model";
+}
+
+export function buildChannelModelNotFoundText(channelId: string): string {
+  return `Model not found. Use ${modelCommandPrefix(channelId)} list to see available models.`;
 }
 
 export function buildChannelCurrentModelMessage(
@@ -617,169 +617,6 @@ export function buildChannelCurrentModelMessage(
     `${displayName} current ${scope} model: ${params.modelLabel}${handleText}.`,
     `Use ${switchCommand} list to see available models, or ${switchCommand} <handle-or-id> to switch.`,
   ].join("\n");
-}
-
-type SlackModelOption = {
-  text: { type: "plain_text"; text: string; emoji?: boolean };
-  value: string;
-  description?: { type: "plain_text"; text: string; emoji?: boolean };
-};
-
-function escapeSlackMrkdwn(text: string): string {
-  return text
-    .replace(/&/g, "&amp;")
-    .replace(/</g, "&lt;")
-    .replace(/>/g, "&gt;");
-}
-
-function truncateSlackPlainText(text: string, limit: number): string {
-  if (text.length <= limit) {
-    return text;
-  }
-  return `${text.slice(0, Math.max(0, limit - 3)).trimEnd()}...`;
-}
-
-function getSlackModelOptionValue(entry: ChannelModelListEntry): string | null {
-  const candidates = [entry.id, entry.handle].filter(
-    (value): value is string => typeof value === "string" && value.length > 0,
-  );
-  return (
-    candidates.find(
-      (value) => value.length <= SLACK_MODEL_OPTION_VALUE_LIMIT,
-    ) ?? null
-  );
-}
-
-function buildSlackModelOption(
-  entry: ChannelModelListEntry,
-): SlackModelOption | null {
-  const value = getSlackModelOptionValue(entry);
-  if (!value) {
-    return null;
-  }
-
-  const labelText =
-    entry.handle === entry.label
-      ? entry.label
-      : `${entry.label} - ${entry.handle}`;
-  const option: SlackModelOption = {
-    text: {
-      type: "plain_text",
-      text: truncateSlackPlainText(labelText, SLACK_MODEL_OPTION_TEXT_LIMIT),
-      emoji: true,
-    },
-    value,
-  };
-  if (entry.description) {
-    option.description = {
-      type: "plain_text",
-      text: truncateSlackPlainText(
-        entry.description,
-        SLACK_MODEL_OPTION_TEXT_LIMIT,
-      ),
-      emoji: true,
-    };
-  }
-  return option;
-}
-
-export function buildSlackModelPickerBlocks(params: {
-  current: {
-    modelLabel: string;
-    modelHandle: string | null;
-    scope?: "agent" | "conversation";
-  };
-  entries: ListModelsResponseModelEntry[];
-  availableHandles?: string[] | null;
-  recentHandles?: string[];
-}): unknown[] | undefined {
-  const entries = params.entries as ChannelModelListEntry[];
-  const byHandle = buildModelEntriesByHandle(entries);
-  const availableHandleList = Array.isArray(params.availableHandles)
-    ? params.availableHandles
-    : null;
-  const availableSet = availableHandleList
-    ? new Set(availableHandleList)
-    : null;
-  const recentEntries = resolveModelHandles({
-    handles: params.recentHandles ?? [],
-    byHandle,
-    availableHandles: availableSet,
-  });
-  const availableEntries = availableHandleList
-    ? resolveModelHandles({ handles: availableHandleList, byHandle })
-    : getFallbackModelEntries(byHandle);
-  const optionEntries = [...recentEntries, ...availableEntries];
-  const options: SlackModelOption[] = [];
-  const seenValues = new Set<string>();
-  for (const entry of optionEntries) {
-    const option = buildSlackModelOption(entry);
-    if (!option || seenValues.has(option.value)) {
-      continue;
-    }
-    seenValues.add(option.value);
-    options.push(option);
-    if (options.length >= SLACK_MODEL_PICKER_OPTION_LIMIT) {
-      break;
-    }
-  }
-
-  if (options.length === 0) {
-    return undefined;
-  }
-
-  const currentScope =
-    params.current.scope === "agent" ? "agent" : "conversation";
-  const currentHandleText = params.current.modelHandle
-    ? `\nHandle: ${escapeSlackMrkdwn(params.current.modelHandle)}`
-    : "";
-  const currentHandle = params.current.modelHandle;
-  const initialOption = options.find((option) => {
-    const entry = optionEntries.find(
-      (candidate) =>
-        candidate.id === option.value || candidate.handle === option.value,
-    );
-    return currentHandle
-      ? entry?.handle === currentHandle || option.value === currentHandle
-      : false;
-  });
-
-  return [
-    {
-      type: "section",
-      text: {
-        type: "mrkdwn",
-        text: `*Current ${currentScope} model:* ${escapeSlackMrkdwn(params.current.modelLabel)}${currentHandleText}`,
-      },
-    },
-    {
-      type: "section",
-      text: {
-        type: "mrkdwn",
-        text: "Choose a model for this routed conversation:",
-      },
-      accessory: {
-        type: "static_select",
-        action_id: SLACK_MODEL_SELECT_ACTION_ID,
-        placeholder: {
-          type: "plain_text",
-          text: "Select a model",
-          emoji: true,
-        },
-        options,
-        ...(initialOption ? { initial_option: initialOption } : {}),
-      },
-    },
-    {
-      type: "context",
-      elements: [
-        {
-          type: "mrkdwn",
-          text: "Need another handle? Mention the app with `@agent /model <handle-or-id>`.",
-        },
-      ],
-    },
-  ];
 }
 
 function formatChannelModelEntry(
@@ -963,7 +800,7 @@ async function handleScopedCommand(params: {
   }
   return {
     text,
-    ...(result.slackBlocks ? { slackBlocks: result.slackBlocks } : {}),
+    ...(result.modelPicker ? { modelPicker: result.modelPicker } : {}),
   };
 }
 
@@ -1094,11 +931,11 @@ export async function tryHandleChannelSlashCommand(
   await adapter.sendDirectReply(
     msg.chatId,
     reply.text,
-    msg.messageId || msg.threadId || reply.slackBlocks
+    msg.messageId || msg.threadId || reply.modelPicker
       ? {
           replyToMessageId: msg.messageId,
           threadId: msg.threadId ?? null,
-          ...(reply.slackBlocks ? { slackBlocks: reply.slackBlocks } : {}),
+          ...(reply.modelPicker ? { modelPicker: reply.modelPicker } : {}),
         }
       : undefined,
   );
